refactored to server cache and created new table for user activity
