+++ backend/helpers/search-helper-functions.js
