+++ backend/routes/products.js
@@ -1,12 +1,11 @@
 const express = require('express')
 const { PrismaClient } = require('../generated/prisma/index.js')
 const { SkinTypes, SkinConcerns, ProductTypes } = require('../enums.js')
-const {updateProductsWithScore} = require('./scoring-helper-functions.js');
-const {cleanSearchQuery} = require('./search-helper-functions.js');
+const {updateProductsWithScore} = require('../helpers/scoring-helper-functions.js');
+const {cleanSearchQuery} = require('../helpers/search-helper-functions.js');
 const prisma = new PrismaClient()
 const router = express.Router()
 
-// http://localhost:3000/products
 router.get('/products', async (req, res) => {
     const page = req.query.page ? parseInt(req.query.page) : 1; // default to page 1
     const limit = req.query.limit ? parseInt(req.query.limit) : 10; //default to limit 10 products per page
@@ -111,7 +110,6 @@ router.get('/products', async (req, res) => {
     });
 });
 
-// http://localhost:3000/change-product-image
 router.put('/change-product-image/:productId', async (req, res) => {
     const {image} = req.body
     const id = parseInt(req.params.productId)
@@ -377,4 +375,5 @@ router.post('/log-interaction/:productId', async (req, res) => {
     }
 
 });
+
 module.exports = router;
